requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.1.0
pyotp==2.9.0
qrcode==8.0
sqlparse==0.5.3

# Testing packages
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==24.1.0
